pyproject.toml
[project]
name = "my-docs"
version = "0.1.0"
description = "Documentation project using MkDocs Material"
requires-python = ">=3.8"

dependencies = [
    "mkdocs",
    "mkdocs-material",
    "mkdocs-minify-plugin",
    "pymdown-extensions",
    "mike",
    "mkdocs-glightbox",
    "mkdocs-localsearch",
    "mkdocs-macros-plugin",
    "pyyaml"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []  # Don't auto-discover Python modules

[tool.setuptools.packages]
find = {}  # Don't auto-discover packages